+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Update Bitmovin's native iOS SDK version to `3.113.0`
 - Update Bitmovin's native Android SDK version to `3.152.0+jason`
 
 ## [1.18.0] - 2026-05-01